requirements.txt
numpy==1.26.4
matplotlib==3.8.2
Pillow==10.2.0
streamlit==1.31.0


